Explicitly set architectures to build for. armv6 and armv7 in Release, and i386 also in Debug
We were having problems with building our iPhone app, as the module was
only configured to compile with armv7, whereas the app is built for both
armv6 and armv7.  Changed build.py to use Debug configuration when
building for the simulator, in the hope that i386 will not be bundled
with the final app.